GitHub Actions: CodeQL & Security Scan (#4)
Adding the following security validations to the build GitHub Action, to run on every commit and every PR:
- CodeQL
- Security Scan

CodeQL runs as a discrete job as it takes a relatively long time to run (approximately 10 minutes). In contrast, Security Scan runs in the existing validation job. This division helps to keep the overall running time down, while minimizing GitHub Action costs.