Updating namespace for publishing to NuGet (#12)
Updating the namespace from NuGet.TransitiveDependency.Finder to NuGetTransitiveDependencyFinder to avoid namespace conflicts when publishing to NuGet.org. NuGet. is a namespace reserved for exclusive use by Microsoft and therefore cannot be published to.

This update also includes a fix for the README.md file, to restore some information previously deleted by accident.