v2.0.0: Addition of '--all' mode for outputting all dependencies (#14)
Updating the NuGet Transitive Dependency Finder to include a new --all mode that outputs all dependencies, irrespective of whether they are transitive or not. This will help to detect whether removal of a dependency could result in a regression due to the change in version of a non-transitive dependency.

README.md has also been updated to reflect the new functionality and the expected mode of operation.

This PR also bumps the release number to v2.0.0, as it constitutes a major release.