.NET 9 – Code Modernization (#633)
# Pull Request

## Summary

This includes changes to improve the handling of regular expressions in
unit tests by converting methods to properties. The most important
changes involve modifying the `FilterRegex` and `MatchingProjectsRegex`
methods to be properties and updating the relevant test cases
accordingly.

Improvements to regular expression handling:

*
[`src/UnitTests/NuGetTransitiveDependencyFinder.UnitTests.ConsoleApp/Process/ProgramRunnerUnitTests.cs`](diffhunk://#diff-859b57afdfdd5c40a9e56cbf1a49b66c88b8e246d86e273414f01b34dca431dbR46-R51):
Converted the `FilterRegex` method to a static partial property and
updated its usage in the `Run_Called_PerformsExpectedActions` test
method.
[[1]](diffhunk://#diff-859b57afdfdd5c40a9e56cbf1a49b66c88b8e246d86e273414f01b34dca431dbR46-R51)
[[2]](diffhunk://#diff-859b57afdfdd5c40a9e56cbf1a49b66c88b8e246d86e273414f01b34dca431dbL62-R70)
[[3]](diffhunk://#diff-859b57afdfdd5c40a9e56cbf1a49b66c88b8e246d86e273414f01b34dca431dbL86-L97)
*
[`src/UnitTests/NuGetTransitiveDependencyFinder.UnitTests/ProjectAnalysis/DependencyFinderTests.cs`](diffhunk://#diff-c751ee84f43e85dab8f6ff3a323356551c97144eba7fea30cbd106354e7db8d3L273-R273):
Converted the `MatchingProjectsRegex` method to a static partial
property and updated its usage in the
`Run_WithNoMatchingProjectDependencies_ReturnsEmptyProjects` and
`Run_WithMatchingProjectDependenciesButNoTransitiveDependencies_ReturnsEmptyProjects`
test methods.
[[1]](diffhunk://#diff-c751ee84f43e85dab8f6ff3a323356551c97144eba7fea30cbd106354e7db8d3L273-R273)
[[2]](diffhunk://#diff-c751ee84f43e85dab8f6ff3a323356551c97144eba7fea30cbd106354e7db8d3L341-R352)